Improve `icmaa-config` module to extend configs on build
