+++ core/scripts/generate-files.ts
@@ -2,6 +2,10 @@ import fs from 'fs';
 import path from 'path';
 import config from 'config';
 
+import extendConfigs from 'icmaa-config/extendConfigs'
+
+extendConfigs()
+
 fs.writeFileSync(
   path.resolve(__dirname, '../build/config.json'),
   JSON.stringify(config)
