--- src/modules/icmaa-config/store/mutation-types.ts
@@ -1,1 +1,0 @@
-export const ICMAA_CONFIG_ADD_MAPS = 'ADD_MAPS'
